refactor(seo): metadados do card saem de um módulo sem next/og
Para reaproveitar o texto do `alt`, o `seo.ts` passou a importar
`@/app/opengraph-image` — um módulo de ROTA do App Router. A aresta é
`lib → app`, ao contrário da direção normal: faz o módulo da rota e os
imports dele rodarem toda vez que `pageMetadata` é importado, e deixa um
ciclo a um import de distância.

`src/lib/og-meta.ts` guarda o que os METADADOS precisam saber sobre o card
(tamanho, content type e o texto do alt) sem nada que o desenhe: não entra
`next/og` nem `ImageResponse`. O `og.tsx` reexporta as constantes para os
três `opengraph-image.tsx` seguirem importando de um lugar só.

Saída conferida byte a byte: `og:image:alt` continua idêntico na home e
nos tópicos, com os 47 tópicos vindos do `roadmap.ts`.
